credit/requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
json5==0.10.0
langchain-groq==0.1.5
langchain-core==0.2.10
langdetect==1.0.9
lxml==5.3.0
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
email-validator==2.1.0
gunicorn==21.2.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
dnspython==2.4.2
selenium
webdriver-manager
cloudscraper
APScheduler>=3.10.0
tabulate>=0.9.0
easyocr
deep-translator
opencv-python-headless
pillow
torch
torchvision
torchaudio
